Add files via upload
I had problems with part 3 and 4, but was able to finish it with the help of my tutor. As I did not unstained how to put a calculation into a function, and if I'm being honest I still don't completely get it. In part 4, I didn't know I had to get() to get the loan_price from the list of dictionaries.
I haven't finished part 5 as I don't understand csv files, or I don't know what the instructions are asking, or I don't understand how to code it in a why the program will understand. I went back to old lessons, and watched videos, and read the links I was referred too, but I don't know how reading the document was going to help me if I didn't understand what I was learning.